Make album thumbnails consistent across page loads
1. Improves performance via caching.
2. Makes it easier to visually navigate when the images stay the same.

Since photos within albums are ordered oldest to newest but albums are
ordered newest to oldest, there's a bit of a tradeoff implemented for
albums with no photos in them, just subalbums.

If the album uses the thumb of the oldest subalbum, it would be unlikely
to change, aiding navigation. However, this means that the subalbum that
the thumb was pulled from would be listed last in the album. If there
are lots of subalbums, it could even be offscreen. This is inconsistent
with how thumbs for albums with photos in them are done (thumb is always
the first image) and just looks sloppy IMO.

The solution that I went with was using the newest subalbum's thumbnail.
This is consistent with photo-filled albums since the thumb will always
be first. The tradeoff is that the thumbnail of the album will change
whenever a new subalbum is added. However, since subalbums will probably
not be added at a high frequency, the thumb should stay constant enough
that it will still aid visual navigation.

Fixes #2