Improve experience for slower connections
Reduced the JPEG quality from 99 -> 85 to bag some significant space
savings with minimal quality loss.

Photo and album thumbnails still kick off loading when they're created,
but only swap the placeholder image for the real once once the image has
finished loading in. The previous behaviour was to swap the placeholder
for the image immediately, regardless of if the image was loaded or not.

On slow connections this makes it easier to see how many albums/photos
exist even if they're taking a long time to load.

Existing issues:
 - There is a bit of a flash of the placeholder when loading a page full
   of images. Changing the placeholder to something more discreet (a
   transparent loading animation?) would go a long way to making this
   problem go away.
 - When an image fails to load for an album thumbnail, the placeholder
   image will never be replaced.
   - This is better than before (no image at all), but it would be
     better if it automatically failed over to the next available
     thumbnail instead.
   - In this case the failed image is also removed from the album so
     the browser won't attempt to load it again if you click the album.
   - I experimented with calling `setTimeout(showAlbum, 1, true)` after
     detecting a failed image to make the albums be laid out again.
     Because the failed image was removed, it would pick up the next
     image and load it. However, the performance penalty of laying out
     all the albums again was pretty bad so I didn't end up committing
     it.
   - This is an edge case anyway...