Change the placeholder image to a loading spinner gif
It blends in a little better and signals the user that something is
happening. It also reduces the flickering effect on page loads since
it's mostly transparent.

The loading spinner is now also used for loading the higher-resolution
images. The 150x150 gif is a little pixellated at that size but using a
higher-res version would kinda miss the point.

Eventually the plan would be to replace the gif with a JS-based spinner
but this is a drop-in 90% solution that works for now.