Don't fall back to file modified time if no exif date is provided
Typically the file modified time doesn't represent when the picture was
taken at all and causes albums to be sorted incorrectly. By always
treating photos with no exif dates data as having a date of
`datetime.min`, they never cause albums to be treated as more recent
than they actually are.

The tradeoff here is that if the file modification time is actually
accurate (is it ever?), it's tossed. Also, this change means the album
thumbnail is usually going to be of an undated picture instead of the
actual first image in the series.