Fix locale encoding issues with non-ascii characters
The default locale of many Linux distributions is "C", which uses an
encoding of "US-ASCII". This is especially true in VMs or containers
since their configuration is usually very basic. For Windows, the
default is "cp1252" (not much better).

In both of these cases, if the scanner attempts to process an image with
characters in the filename or metadata that can't be represented in that
charset (a LOT of special characters), it will raise encoding errors.
For better traceability, extra logging around these errors was added.

To fix this, the scanner will now always use UTF-8 when communicating
with an `exiftool` subprocess. This should always work since even on
systems that use other encodings, `exiftool` will output text as
utf8-encoded bytes[0].

The one issue with this is that in Python 3.5, the `subprocess`
functions don't take an `encoding` parameter, instead using whatever
`locale.getpreferredencoding(False)` returns. To work around this, if
the scanner is run using Python 3.5 with a locale that doesn't specify
an encoding of "UTF-8", it will attempt to set the locale's text
encoding (`LC_CTYPE`) to UTF-8.

[0]: https://sno.phy.queensu.ca/~phil/exiftool/faq.html#Q10